refactor: modularize deep research agent
