+++ apps/agent/src/agents/deep-research/config/index.ts
@@ -1,0 +1,2 @@
+export * from "./defaults.js";
+export * from "./subagents.js";
