+++ apps/agent/src/agents/deep-research/middleware/shared/reserved-tools.ts
@@ -1,0 +1,8 @@
+export const RESERVED_TOOL_NAMES = new Set([
+  "ls",
+  "read_file",
+  "write_file",
+  "edit_file",
+  "write_todos",
+  "task",
+]);
