+++ apps/agent/src/agents/deep-research/prompts/index.ts
@@ -1,0 +1,4 @@
+export * from "./agent.js";
+export * from "./critique-subagent.js";
+export * from "./planner-subagent.js";
+export * from "./research-subagent.js";
